Remove seek commands
Seek and track switching commands need to be enabled at the same time to be available (see https://stackoverflow.com/a/24818340/760435). If no next track is available, for example, seek forward will not be available.

Moreover, skips take precedence over track change and seeks. Since we now support a seek bar in the control center, I suggest we remove seek button support. This is difficult to document and tricky to get right.